+++ examples/commonjs/App.vue
@@ -1,18 +1,16 @@
 <template>
-  <div style="background: #000;padding:50px;">
+  <div style="background: #fff;padding:50px;">
     <star-rating
       v-model:rating="rating"
       :round-start-rating="false"
       :star-points="[23,2, 14,17, 0,19, 10,34, 7,50, 23,43, 38,50, 36,34, 46,19, 31,17]"
-      :border-width="3"
-      border-color="#fff"
-      :padding="5"
+      :border-width="5"
+      border-color="black"
+      active-border-color="#ccc"
       :clearable="true"
       inactive-color="#fff"
-      :apply-active-color-on-hover="true"
+      :apply-active-color-on-hover="false"
       :animate="true"
-      glow-color="#fff"
-      :glow="5"
     />
     <a
       href="#"